requirements.txt
# Core dependencies
rasterio>=1.3.0
GDAL>=3.6.0
numpy>=1.24.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
scikit-image>=0.21.0
opencv-python>=4.8.0

# Geospatial and scientific computing
matplotlib>=3.7.0
scipy>=1.11.0
geopandas>=0.14.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# SMS alert system
twilio>=8.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing utilities
tqdm>=4.65.0
pyproj>=3.6.0
shapely>=2.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0